+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## Gript 0.0.39 (24-03-2016)
 
- - proxy port can now be customized in config. Proxy options like `url` and `hostHeader` moved to the `proxy` section of the config
+ - proxy port can now be customized in the config.
 
 ## Gript 0.0.38 (23-03-2016)